+++ src/components/Note/NoteMain/NoteTitle/NoteTitle.tsx
@@ -1,38 +1,18 @@
 import calendarBlack from '@/assets/webps/Note/calendarBlack.webp';
 import timeBlack from '@/assets/webps/Note/timeBlack.webp';
-import link from '@/assets/svgs/Note/link.svg';
 import mic from '@/assets/svgs/Note/mic.svg';
 import headphones from '@/assets/svgs/Note/headphones.svg';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { NoteDetail } from '@/pages/Note/dto';
 import { PlatformModal } from '@/components/Note/NoteMain/NoteTitle/PlatformModal/PlatformModal';
+import CopyLink from '@/components/Note/NoteMain/NoteTitle/CopyLink/CopyLink';
 
 interface NoteTitleProps {
   noteData: NoteDetail | null;
 }
 
 export const NoteTitle: React.FC<NoteTitleProps> = ({ noteData }) => {
-  const [isCopied, setIsCopied] = useState(false);
   const [isPlatformModalOpen, setIsPlatformModalOpen] = useState(false);
-
-  // 링크 복사 함수
-  const handleCopyLink = async () => {
-    try {
-      const currentUrl = window.location.href;
-      await navigator.clipboard.writeText(currentUrl); // ClipBoard API 사용 - 현재 페이지 URL 복사
-      setIsCopied(true); // 복사 성공 상태 업데이트
-    } catch (error) {
-      console.error('링크 복사 실패:', error);
-    }
-  };
-
-  // isCopied 상태 변경 감지
-  useEffect(() => {
-    if (isCopied) {
-      alert('링크가 복사되었습니다!');
-      setTimeout(() => setIsCopied(false), 2000); // 2초 후 상태 초기화
-    }
-  }, [isCopied]);
 
   if (!noteData) {
     return <p>노트 정보를 불러오는 중입니다...</p>;
@@ -40,6 +20,7 @@ export const NoteTitle: React.FC<NoteTitleProps> = ({ noteData }) => {
 
   const date = new Date(noteData.createdAt);
   const formattedDate = `${date.getFullYear().toString().slice(-2)}.${(date.getMonth() + 1).toString().padStart(2, '0')}.${date.getDate().toString().padStart(2, '0')}`;
+  const currentUrl = window.location.href;
 
   return (
     <>
@@ -47,13 +28,7 @@ export const NoteTitle: React.FC<NoteTitleProps> = ({ noteData }) => {
         <div className="flex justify-between items-center mt-[30px]">
           <p className="text-[26px] font-bold ml-12 mr-[120px] leading-[30px]">{noteData.name}</p>
           <div className="flex text-white text-[10.5px] leading-[18px]">
-            <div
-              onClick={handleCopyLink}
-              className="flex items-center w-[88px] h-[30px] mr-3 bg-mainBlack rounded-[3.2px] cursor-pointer"
-            >
-              <img className="w-[15px] h-[15px] ml-3 mr-[6px]" src={link} alt="link" />
-              <span className="font-[500]">링크 복사</span>
-            </div>
+            <CopyLink url={currentUrl} />
             <div
               onClick={() => {
                 setIsPlatformModalOpen(true);
